RawIntensity released functions work internally with floats instead of doubles now.

+++ IntensityAttributingKit/RawIntensityMapping.swift
@@ -8,11 +8,11 @@
 
 import Foundation
 
-typealias RawIntensityMappingFunction = ((raw:Double)->Int)
+typealias RawIntensityMappingFunction = ((raw:Float)->Int)
 
 public enum RawIntensityMapping {
-    case Linear(threshold:Double,ceiling:Double),
-    LogAx(a:Double,threshold:Double,ceiling:Double)
+    case Linear(threshold:Float,ceiling:Float),
+    LogAx(a:Float,threshold:Float,ceiling:Float)
     
     var function:RawIntensityMappingFunction{
         switch self{
@@ -39,13 +39,13 @@ public enum RawIntensityMapping {
     init!(dictDescription:[String:AnyObject]){
         //require matching name, otherwise fill in default values where needed
         guard let name = dictDescription["name"] as? String else {return nil}
-        let threshold = (dictDescription["threshold"] as? Double) ?? 0.0
-        let ceiling = (dictDescription["ceiling"] as? Double) ?? 1.0
+        let threshold = (dictDescription["threshold"] as? Float) ?? 0.0
+        let ceiling = (dictDescription["ceiling"] as? Float) ?? 1.0
         
         switch name {
             case "Linear": self = .Linear(threshold: threshold, ceiling: ceiling)
             case "LogAx":
-                let a = (dictDescription["a"] as? Double) ?? 10.0
+                let a = (dictDescription["a"] as? Float) ?? 10.0
                 self = .LogAx(a: a, threshold: threshold, ceiling: ceiling)
         default:
             return nil
@@ -53,13 +53,13 @@ public enum RawIntensityMapping {
         
     }
     
-    var threshold:Double {
+    var threshold:Float {
         switch self {
         case .Linear(let threshold, _): return threshold
         case .LogAx(_, let threshold, _): return threshold
         }
     }
-    var ceiling:Double {
+    var ceiling:Float {
         switch self {
         case .Linear( _, let ceiling): return ceiling
         case .LogAx( _,  _, let ceiling): return ceiling
@@ -67,7 +67,7 @@ public enum RawIntensityMapping {
     }
     
     ///Parameter "a", where appropriate
-    var a:Double? {
+    var a:Float? {
         switch self {
         case .Linear( _, _): return nil
         case .LogAx( let a,  _, _): return a
@@ -76,8 +76,8 @@ public enum RawIntensityMapping {
 }
 
 private struct LinearMapping{
-    static func generateFunction(threshold thresh:Double,ceiling:Double)->RawIntensityMappingFunction{
-        func linearMapping(raw:Double)->Int{
+    static func generateFunction(threshold thresh:Float,ceiling:Float)->RawIntensityMappingFunction{
+        func linearMapping(raw:Float)->Int{
             guard raw > thresh else {return 0}
             guard raw <= ceiling else {return 100}
             return Int(100 * (raw - thresh) / (ceiling - thresh))
@@ -87,9 +87,9 @@ private struct LinearMapping{
 }
 
 private struct LogAxMapping {
-    static func generateFunction(a:Double, threshold:Double, ceiling:Double)->RawIntensityMappingFunction{
+    static func generateFunction(a:Float, threshold:Float, ceiling:Float)->RawIntensityMappingFunction{
         let factor = log(1 + (ceiling - threshold)*a) / 100
-        func rim(raw:Double)->Int{
+        func rim(raw:Float)->Int{
             guard raw > threshold else {return 0}
             guard raw <= ceiling else {return 100}
             return Int((log(1 + (raw - threshold)*a) ) / factor)
