+++ IntensityAttributingKit/TouchInterpreters.swift
@@ -30,53 +30,53 @@ public enum IATouchInterpreter:String{
 
 
 protocol IATouchInterpretingProtocol {
-    mutating func updateIntensityYieldingRawResult(withTouch touch:UITouch)->Double
-    mutating func endInteractionYieldingRawResult(withTouch touch:UITouch?)->Double
+    mutating func updateIntensityYieldingRawResult(withTouch touch:UITouch)->Float
+    mutating func endInteractionYieldingRawResult(withTouch touch:UITouch?)->Float
     mutating func cancelInteraction()
-    var currentRawValue:Double! {get}
+    var currentRawValue:Float! {get}
 }
 
 
 private struct ForceTouchInterpreter:IATouchInterpretingProtocol {
     
     static var ftOption:FTOptions = .SmoothedLastTen
-    static var maximumPossibleForce:Double = 6.66667  //this should be determined by device type
+    static var maximumPossibleForce:Float = 6.66667  //this should be determined by device type
     
-    var history:[Double] = []
-    var maxForce:Double = 0.0
+    var history:[Float] = []
+    var maxForce:Float = 0.0
     
-    mutating func updateIntensityYieldingRawResult(withTouch touch:UITouch)->Double{
+    mutating func updateIntensityYieldingRawResult(withTouch touch:UITouch)->Float{
         switch ForceTouchInterpreter.ftOption {
         case .AvgLastTen:
-            history.append(Double(touch.force))
+            history.append(Float(touch.force))
             return avgLastTenPressure
         case .SmoothedLastTen:
-            history.append(Double(touch.force))
+            history.append(Float(touch.force))
             return smoothedLastTen
         case .PeakPressure:
-            maxForce = max(maxForce, Double(touch.force))
+            maxForce = max(maxForce, Float(touch.force))
             return peakScaledPressure
         }
     }
     
-    mutating func endInteractionYieldingRawResult(withTouch touch:UITouch?)->Double{
-        var result:Double!
+    mutating func endInteractionYieldingRawResult(withTouch touch:UITouch?)->Float{
+        var result:Float!
         switch ForceTouchInterpreter.ftOption {
         case .AvgLastTen:
             if let touch = touch {
-                history.append(Double(touch.force))
+                history.append(Float(touch.force))
             }
             result = avgLastTenPressure
             history.removeAll()
         case .SmoothedLastTen:
             if let touch = touch {
-                history.append(Double(touch.force))
+                history.append(Float(touch.force))
             }
             result = smoothedLastTen
             history.removeAll()
         case .PeakPressure:
             if let touch = touch {
-                maxForce = max(maxForce, Double(touch.force))
+                maxForce = max(maxForce, Float(touch.force))
             }
             result = peakScaledPressure
             maxForce = 0.0
@@ -89,7 +89,7 @@ private struct ForceTouchInterpreter:IATouchInterpretingProtocol {
         maxForce = 0.0
     }
     
-    var currentRawValue:Double! {
+    var currentRawValue:Float! {
         switch ForceTouchInterpreter.ftOption {
         case .AvgLastTen: return avgLastTenPressure
         case .SmoothedLastTen: return smoothedLastTen
@@ -105,31 +105,31 @@ private struct ForceTouchInterpreter:IATouchInterpretingProtocol {
     }
     
     
-    private var avgLastTenPressure:Double {
+    private var avgLastTenPressure:Float {
         let count = history.count
         guard count > 1 else {return 0.0}
         if count < 10 {
-            return (history[1..<count].reduce(0.0, combine: +) / Double(count - 1)) / ForceTouchInterpreter.maximumPossibleForce
+            return (history[1..<count].reduce(0.0, combine: +) / Float(count - 1)) / ForceTouchInterpreter.maximumPossibleForce
         } else {
-            return (history[(count - 10)..<count].reduce(0.0, combine: +) / Double(10)) / ForceTouchInterpreter.maximumPossibleForce
+            return (history[(count - 10)..<count].reduce(0.0, combine: +) / Float(10)) / ForceTouchInterpreter.maximumPossibleForce
         }
         
     }
     
-    private var peakScaledPressure:Double {
+    private var peakScaledPressure:Float {
         return maxForce / ForceTouchInterpreter.maximumPossibleForce
     }
  
     
-    private var smoothedLastTen:Double {
+    private var smoothedLastTen:Float {
         let count = history.count
         guard count > 2 else {return history.maxElement() ?? 0.0}
-        //var truncated:[Double]!
+        //var truncated:[Float]!
         
-        var sum:Double = 0
-        var max:Double = 0
-        var min:Double = 1
-        var adjustedCount:Double!
+        var sum:Float = 0
+        var max:Float = 0
+        var min:Float = 1
+        var adjustedCount:Float!
         
         if count < 10 {
             for val in history[0..<count] {
@@ -137,7 +137,7 @@ private struct ForceTouchInterpreter:IATouchInterpretingProtocol {
                 if val > max {max = val}
                 if val < min {min = val}
             }
-            adjustedCount = Double(count - 2)
+            adjustedCount = Float(count - 2)
         } else {
             for val in history[(count - 10)..<count] {
                 sum += val
@@ -150,14 +150,14 @@ private struct ForceTouchInterpreter:IATouchInterpretingProtocol {
         return ((sum - max - min) / adjustedCount) / ForceTouchInterpreter.maximumPossibleForce
         
 //        if count < 10 {
-//            truncated = Array<Double>(history[1..<count])
+//            truncated = Array<Float>(history[1..<count])
 //        } else {
-//            truncated = Array<Double>(history[(count - 10)..<count])
+//            truncated = Array<Float>(history[(count - 10)..<count])
 //        }
 //        truncated.sortInPlace()
 //        truncated.removeLast()
 //        truncated.removeFirst()
-//        return (truncated.reduce(0.0, combine: +) / Double(truncated.count)) / ForceTouchInterpreter.maximumPossibleForce
+//        return (truncated.reduce(0.0, combine: +) / Float(truncated.count)) / ForceTouchInterpreter.maximumPossibleForce
     }
     
 }
@@ -165,19 +165,19 @@ private struct ForceTouchInterpreter:IATouchInterpretingProtocol {
 
 
 private struct DurationTouchInterpreter:IATouchInterpretingProtocol{
-    static var fullTouchDuration:Double = 0.5
+    static var fullTouchDuration:Float = 0.5
     
     var touchStartTime:CFTimeInterval!
     
-    mutating func updateIntensityYieldingRawResult(withTouch touch:UITouch)->Double{
+    mutating func updateIntensityYieldingRawResult(withTouch touch:UITouch)->Float{
         if touchStartTime == nil {
             touchStartTime = CACurrentMediaTime()
             return 0.0
         } else {
-            return max((CACurrentMediaTime() - touchStartTime) / DurationTouchInterpreter.fullTouchDuration, 1.0)
+            return max(Float(CACurrentMediaTime() - touchStartTime) / DurationTouchInterpreter.fullTouchDuration, 1.0)
         }
     }
-    mutating func endInteractionYieldingRawResult(withTouch touch:UITouch?)->Double{
+    mutating func endInteractionYieldingRawResult(withTouch touch:UITouch?)->Float{
         let raw = self.currentRawValue ?? 0.0
         touchStartTime = nil
         return raw
@@ -185,8 +185,8 @@ private struct DurationTouchInterpreter:IATouchInterpretingProtocol{
     mutating func cancelInteraction(){
         touchStartTime = nil
     }
-    var currentRawValue:Double! {
+    var currentRawValue:Float! {
         guard touchStartTime != nil else {return nil}
-        return min((CACurrentMediaTime() - touchStartTime) / DurationTouchInterpreter.fullTouchDuration, 1.0)
+        return min(Float(CACurrentMediaTime() - touchStartTime) / DurationTouchInterpreter.fullTouchDuration, 1.0)
     }
 }
